+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add SCIP function SCIprowGetOriginCons
 - Add getConsOriginConshdlrtype to Row
 - Add possibility to use sine and cosing
+- Add ability to set priced variable score 
 
 ### Fixed
 ### Changed